+++ test/SingleParameterCacheKey.js
@@ -1,5 +1,6 @@
 // test
 import test from 'ava';
+import _ from 'lodash';
 
 // src
 import SingleParameterCacheKey from 'src/SingleParameterCacheKey';
@@ -50,6 +51,45 @@ test('if matches will return true if the key passed is not a multi-parameter key
   const isMultiParamKey = false;
 
   const result = cacheKey.matches(newKey, isMultiParamKey);
+
+  t.true(result);
+});
+
+test('if matchesCustom will return false if the key passed is a multi-parameter key', (t) => {
+  const existingKey = [{foo: 'foo'}];
+
+  const cacheKey = new SingleParameterCacheKey(existingKey);
+
+  const newKey = [{foo: 'foo'}, {bar: 'bar'}];
+  const isMultiParamKey = true;
+
+  const result = cacheKey.matchesCustom(newKey, isMultiParamKey, _.isEqual);
+
+  t.false(result);
+});
+
+test('if matchesCustom will return false if the key passed is not a multi-parameter key that is not equal based on the custom method', (t) => {
+  const existingKey = [{foo: 'foo'}];
+
+  const cacheKey = new SingleParameterCacheKey(existingKey);
+
+  const newKey = [{foo: 'bar'}];
+  const isMultiParamKey = false;
+
+  const result = cacheKey.matchesCustom(newKey, isMultiParamKey, _.isEqual);
+
+  t.false(result);
+});
+
+test('if matchesCustom will return true if the key passed is not a multi-parameter key that is equal based on the custom method', (t) => {
+  const existingKey = [{foo: 'foo'}];
+
+  const cacheKey = new SingleParameterCacheKey(existingKey);
+
+  const newKey = [{foo: 'foo'}];
+  const isMultiParamKey = false;
+
+  const result = cacheKey.matchesCustom(newKey, isMultiParamKey, _.isEqual);
 
   t.true(result);
 });